pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "inspec"
version = "0.5"
description = "Inspect audio and image files from command line"
readme = "README.md"
authors = [{name = "Kevin Yu", email = "thekevinyu@gmail.edu"}]
license = {file = "LICENSE"}
keywords = ["spectrogram", "audio", "visualization", "sound", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.0",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "Pillow",
    "click",
    "numpy",
    "opencv-python",
    "pydantic",
    "sounddevice",
    "soundfile",
    "typing-extensions",
    "wheel",
    "windows-curses; platform_system=='Windows'",
]

[project.optional-dependencies]
dev = [
    "black",
    "build",
    "flake8",
    "ipdb",
    "isort",
    "pre-commit",
    "pyright",
    "pytest",
    "twine",
]

[project.urls]
Repository = "https://github.com/kevinyu/inspec"

[project.scripts]
inspec = "inspec_core.cli:cli"

[tool.setuptools]
py-modules = ["inspec"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["inspec_core", "inspec_core*"]
exclude = ["*/examples/*", "*/tests/*"]
namespaces = false

[tool.isort]
profile = "black"
src_paths = "src"
extend_skip = [".env", "env"]

[tool.flake8]
ignore = "E203, E266, E501, W503, F403, F401"
max-line-length = 100
max-complexity = 18
select = "B,C,E,F,W,T4,B9"

[tool.pyright]
venvPath = "."
venv = "env"
